# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Desertica API
  version: 0.1.0
  summary: Contrato del backend transaccional de Desertica.
  description: |
    Fuente de verdad del contrato entre `desertica-api`, `desertica-web` y `desertica-backoffice`.
    Un cambio de contrato se hace primero en este archivo, por PR, y después se implementa.

    ## Convenciones
    - **Dinero**: enteros en la unidad menor (`*Cents`), con IGV incluido. El precio es único para todos
      los clientes. Monedas: `USD` y `PEN`.
    - **Fechas**: ISO 8601 en UTC. Las reglas de negocio (cortes de venta, recordatorios) usan `America/Lima`.
    - **Paginación**: `page` (desde 1) y `pageSize` (máximo 100). Las listas responden `{ data, meta }`.
    - **Errores**: `{ statusCode, error, message, details? }`. Los errores transversales no se repiten en cada
      operación: `400` (JSON mal formado), `401`, `403` (con `details.required` = permisos que faltan),
      `422` (validación de la entrada, incluidos ids mal formados; `message` lista cada campo), `429` y `5xx`.
    - **Idempotencia**: los `POST` que crean reservas y pagos aceptan `Idempotency-Key`; repetir la clave
      devuelve la respuesta original. Reusar la clave con otro cuerpo responde `422`. Una petición simultánea
      con la misma clave espera a la primera. En `createPublicBooking` la repetición devuelve la misma reserva
      con un `accessToken` nuevo (el original no se guarda en claro).
...
  license:
    name: Proprietary
...
servers:
  - url: /api
...
security:
  - bearerAuth: []
paths:
...
  /public/holds:
    post:
      operationId: createHold
      tags: [Public]
      summary: Bloquear cupo temporalmente
      description: Reserva los asientos durante unos minutos mientras el cliente completa el pago.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [departureId, seats]
              properties:
                departureId: { type: string, format: uuid }
                seats: { type: integer, minimum: 1, maximum: 20 }
                turnstileToken: { type: string }
      responses:
        '404': { $ref: '#/components/responses/Error' }
        '201':
          description: Cupo bloqueado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Hold' }
        '409':
          description: Sin cupo suficiente
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/Error' }
  /public/holds/{token}:
    delete:
      operationId: releaseHold
      tags: [Public]
      summary: Liberar un bloqueo de cupo
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Liberado }
  /public/bookings:
    post:
      operationId: createPublicBooking
      tags: [Public]
      summary: Crear una reserva desde un bloqueo de cupo
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePublicBooking' }
      responses:
        '201':
          description: Reserva creada pendiente de pago
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicBookingCreated' }
        '409': { $ref: '#/components/responses/Error' }
        '410':
          description: El bloqueo de cupo expiró
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /public/bookings/access:
    post:
      operationId: requestBookingAccess
      tags: [Public]
      summary: Pedir acceso a "mi reserva"
      description: Envía por email un enlace con el token de acceso. Responde igual exista o no la reserva.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reference, email]
              properties:
                reference: { type: string }
                email: { type: string, format: email }
      responses:
        '202': { description: Aceptado }
        '429': { $ref: '#/components/responses/Error' }
  /public/bookings/{reference}:
    get:
      operationId: getPublicBooking
      tags: [Public]
      summary: Ver una reserva con el token del cliente
      security:
        - bookingToken: []
      parameters:
        - $ref: '#/components/parameters/Reference'
      responses:
        '200':
          description: Reserva
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicBooking' }
        '401': { $ref: '#/components/responses/Error' }
...
  /public/payment-links/{token}:
    get:
      operationId: getPaymentLinkInfo
      tags: [Public]
      summary: Ver un enlace de pago
      security: []
      parameters:
        - $ref: '#/components/parameters/LinkToken'
      responses:
        '200':
          description: Datos del enlace
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentLinkInfo' }
        '404': { $ref: '#/components/responses/Error' }
        '410': { $ref: '#/components/responses/Error' }
...
  /bookings:
    get:
      operationId: listBookings
      tags: [Bookings]
      summary: Listar reservas
      x-permission: bookings:read
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/BookingStatus' }
        - name: departureId
          in: query
          schema: { type: string, format: uuid }
        - name: tourRefId
          in: query
          schema: { type: string, format: uuid }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: q
          in: query
          description: Busca por referencia, nombre, correo o documento.
          schema: { type: string }
      responses:
        '200':
          description: Reservas
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/BookingSummary' }
                  meta: { $ref: '#/components/schemas/Meta' }
    post:
      operationId: createManualBooking
      tags: [Bookings]
      summary: Crear una reserva manual (asistida)
      description: Para ventas por WhatsApp o en persona. Descuenta cupo sin pasar por el bloqueo temporal.
      x-permission: bookings:write
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateManualBooking' }
      responses:
        '404': { $ref: '#/components/responses/Error' }
        '201':
          description: Reserva
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
        '409': { $ref: '#/components/responses/Error' }
  /bookings/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      operationId: getBooking
      tags: [Bookings]
      summary: Ver una reserva con pagos, comprobantes y descargos
      x-permission: bookings:read
      responses:
        '200':
          description: Reserva
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
        '404': { $ref: '#/components/responses/Error' }
    patch:
      operationId: updateBooking
      tags: [Bookings]
      summary: Editar notas, datos de contacto o pasajeros
      x-permission: bookings:write
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateBooking' }
      responses:
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '200':
          description: Reserva
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    bookingToken:
      type: apiKey
      in: header
      name: X-Booking-Token
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Reference:
      name: reference
      in: path
      required: true
      schema: { type: string, example: DST-7K4Q9M }
...
    LinkToken:
      name: token
      in: path
      required: true
      schema: { type: string }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema: { type: string, maxLength: 128 }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      required: [statusCode, error, message]
      properties:
        statusCode: { type: integer }
        error: { type: string }
        message:
          oneOf:
            - type: string
            - type: array
              items: { type: string }
        details:
          type: object
          additionalProperties: true
    Meta:
      type: object
      required: [page, pageSize, total]
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
    Currency:
      type: string
      enum: [USD, PEN]
    IdDocType:
      type: string
      enum: [DNI, CE, PASSPORT, RUC]
...
    BookingStatus:
      type: string
      enum: [PENDING_PAYMENT, CONFIRMED, COMPLETED, CANCELLED, NO_SHOW]
    PaymentProvider:
      type: string
      enum: [STRIPE, CULQI, MANUAL]
    PaymentMethod:
      type: string
      enum: [CARD, WALLET, YAPE, CASH, TRANSFER, OTHER]
    PaymentKind:
      type: string
      enum: [FULL, DEPOSIT, BALANCE]
    PaymentStatus:
      type: string
      enum: [PENDING, REQUIRES_ACTION, SUCCEEDED, FAILED, CANCELLED, PARTIALLY_REFUNDED, REFUNDED, DISPUTED]
...
    DocumentType:
      type: string
      enum: [BOLETA, FACTURA, NOTA_CREDITO, NOTA_DEBITO]
    DocumentStatus:
      type: string
      enum: [PENDING, ISSUED, ACCEPTED, REJECTED, VOIDED, ERROR]
    WaiverStatus:
      type: string
      enum: [PENDING, SIGNED]
...
    Hold:
      type: object
      required: [token, departureId, seats, expiresAt]
      properties:
        token: { type: string }
        departureId: { type: string, format: uuid }
        seats: { type: integer }
        expiresAt: { type: string, format: date-time }
    Billing:
      type: object
      required: [docType, name, idDocType, idDocNumber]
      description: Datos del receptor del comprobante.
      properties:
        docType: { type: string, enum: [BOLETA, FACTURA] }
        name: { type: string, description: Nombre o razón social. }
        idDocType: { $ref: '#/components/schemas/IdDocType' }
        idDocNumber: { type: string }
        address: { type: string }
        email: { type: string, format: email }
    CustomerInput:
      type: object
      required: [email, firstName, lastName]
      properties:
        email: { type: string, format: email }
        firstName: { type: string }
        lastName: { type: string }
        phone: { type: string }
        country: { type: string, minLength: 2, maxLength: 2 }
        idDocType: { $ref: '#/components/schemas/IdDocType' }
        idDocNumber: { type: string }
        locale: { type: string }
    PassengerInput:
      type: object
      required: [firstName, lastName]
      properties:
        firstName: { type: string }
        lastName: { type: string }
        idDocType: { $ref: '#/components/schemas/IdDocType' }
        idDocNumber: { type: string }
        birthDate: { type: string, format: date }
        nationality: { type: string }
        emergencyContactName: { type: string }
        emergencyContactPhone: { type: string }
    Attribution:
      type: object
      properties:
        utmSource: { type: string }
        utmMedium: { type: string }
        utmCampaign: { type: string }
        utmContent: { type: string }
        utmTerm: { type: string }
        gclid: { type: string }
        fbclid: { type: string }
        landingPath: { type: string }
        referrer: { type: string }
    CreatePublicBooking:
      type: object
      required: [holdToken, currency, adults, customer, billing, paymentKind, acceptedLegalDocumentIds]
      properties:
        holdToken: { type: string }
        currency: { $ref: '#/components/schemas/Currency' }
        adults: { type: integer, minimum: 1, maximum: 20 }
        children: { type: integer, minimum: 0, maximum: 20, default: 0 }
        customer: { $ref: '#/components/schemas/CustomerInput' }
        passengers:
          type: array
          items: { $ref: '#/components/schemas/PassengerInput' }
        billing: { $ref: '#/components/schemas/Billing' }
        paymentKind: { type: string, enum: [FULL, DEPOSIT] }
        acceptedLegalDocumentIds:
          type: array
          description: Versiones de términos, privacidad y política de cancelación aceptadas.
          items: { type: string, format: uuid }
        notes: { type: string, maxLength: 1000 }
        locale: { type: string }
        attribution: { $ref: '#/components/schemas/Attribution' }
        turnstileToken: { type: string }
    CreateManualBooking:
      type: object
      required: [departureId, currency, adults, customer, billing]
      properties:
        departureId: { type: string, format: uuid }
        currency: { $ref: '#/components/schemas/Currency' }
        adults: { type: integer, minimum: 1 }
        children: { type: integer, minimum: 0, default: 0 }
        customer: { $ref: '#/components/schemas/CustomerInput' }
        passengers:
          type: array
          items: { $ref: '#/components/schemas/PassengerInput' }
        billing: { $ref: '#/components/schemas/Billing' }
        depositCents: { type: integer, minimum: 0 }
        overrideTotalCents:
          type: integer
          minimum: 0
          description: Precio acordado distinto del de lista; requiere `bookings:override`.
        notes: { type: string }
        sendConfirmation: { type: boolean, default: true }
    UpdateBooking:
      type: object
      properties:
        notes: { type: string }
        customer: { $ref: '#/components/schemas/CustomerInput' }
        billing: { $ref: '#/components/schemas/Billing' }
        passengers:
          type: array
          items: { $ref: '#/components/schemas/PassengerInput' }
    Passenger:
      allOf:
        - $ref: '#/components/schemas/PassengerInput'
        - type: object
          required: [id]
          properties:
            id: { type: string, format: uuid }
    BookingSummary:
      type: object
      required: [id, reference, status, departureId, startsAt, tourSlug, customerName, currency, totalCents, paidCents]
      properties:
        id: { type: string, format: uuid }
        reference: { type: string }
        status: { $ref: '#/components/schemas/BookingStatus' }
        source: { type: string, enum: [WEB, MANUAL] }
        departureId: { type: string, format: uuid }
        startsAt: { type: string, format: date-time }
        tourSlug: { type: string }
        customerName: { type: string }
        adults: { type: integer }
        children: { type: integer }
        currency: { $ref: '#/components/schemas/Currency' }
        totalCents: { type: integer }
        paidCents: { type: integer }
        createdAt: { type: string, format: date-time }
    Booking:
      allOf:
        - $ref: '#/components/schemas/BookingSummary'
        - type: object
          properties:
            customer: { $ref: '#/components/schemas/Customer' }
            depositCents: { type: [integer, 'null'] }
            refundedCents: { type: integer, description: Devuelto efectivamente por la pasarela o a mano. }
            refundDueCents:
              type: integer
              description: Reembolsos calculados (cancelación, cambio de precio) que aún no se ejecutaron.
            billing: { $ref: '#/components/schemas/Billing' }
            notes: { type: [string, 'null'] }
            locale: { type: string }
            priceSnapshot: { type: object, additionalProperties: true }
            cancellationSnapshot: { type: object, additionalProperties: true }
            attribution: { $ref: '#/components/schemas/Attribution' }
            cancelledAt: { type: [string, 'null'], format: date-time }
            cancelReason: { type: [string, 'null'] }
            passengers:
              type: array
              items: { $ref: '#/components/schemas/Passenger' }
            payments:
              type: array
              items: { $ref: '#/components/schemas/Payment' }
            documents:
              type: array
              items: { $ref: '#/components/schemas/Document' }
            waivers:
              type: array
              items: { $ref: '#/components/schemas/Waiver' }
    PublicBooking:
      type: object
      description: Vista de la reserva para el cliente (sin datos internos).
      required: [reference, status, tourSlug, startsAt, currency, totalCents, paidCents, pendingCents]
      properties:
        reference: { type: string }
        status: { $ref: '#/components/schemas/BookingStatus' }
        tourSlug: { type: string }
        startsAt: { type: string, format: date-time }
        meetingPoint: { type: [string, 'null'] }
        adults: { type: integer }
        children: { type: integer }
        currency: { $ref: '#/components/schemas/Currency' }
        totalCents: { type: integer }
        paidCents: { type: integer }
        pendingCents: { type: integer }
        depositCents: { type: [integer, 'null'] }
        cancellationTiers:
          type: array
          items:
            type: object
            properties:
              hoursBefore: { type: integer }
              refundPercent: { type: integer }
        waivers:
          type: array
          items:
            type: object
            properties:
              token: { type: string }
              passengerName: { type: [string, 'null'] }
              status: { $ref: '#/components/schemas/WaiverStatus' }
        documents:
          type: array
          items:
            type: object
            properties:
              docType: { $ref: '#/components/schemas/DocumentType' }
              number: { type: string, example: B001-00000012 }
              pdfUrl: { type: string, format: uri }
    PublicBookingCreated:
      type: object
      required: [booking, accessToken]
      properties:
        booking: { $ref: '#/components/schemas/PublicBooking' }
        accessToken:
          type: string
          description: Se envía como `X-Booking-Token` en las llamadas siguientes y por email.
        paymentOptions:
          type: array
          description: Pasarelas disponibles para la moneda de la reserva.
          items:
            type: object
            required: [provider, kinds]
            properties:
              provider: { type: string, enum: [STRIPE, CULQI] }
              kinds:
                type: array
                items: { $ref: '#/components/schemas/PaymentKind' }

    # ------------------------------------------------------------- Pagos
...
    PaymentLinkInfo:
      type: object
      required: [reference, kind, currency, amountCents, expiresAt, paymentOptions]
      properties:
        reference: { type: string }
        tourSlug: { type: string }
        startsAt: { type: string, format: date-time }
        kind: { $ref: '#/components/schemas/PaymentKind' }
        currency: { $ref: '#/components/schemas/Currency' }
        amountCents: { type: integer }
        expiresAt: { type: string, format: date-time }
        paymentOptions:
          type: array
          items: { type: string, enum: [STRIPE, CULQI] }
    Payment:
      type: object
      required: [id, bookingId, provider, method, kind, status, currency, amountCents]
      properties:
        id: { type: string, format: uuid }
        bookingId: { type: string, format: uuid }
        bookingReference: { type: string }
        provider: { $ref: '#/components/schemas/PaymentProvider' }
        method: { $ref: '#/components/schemas/PaymentMethod' }
        kind: { $ref: '#/components/schemas/PaymentKind' }
        status: { $ref: '#/components/schemas/PaymentStatus' }
        currency: { $ref: '#/components/schemas/Currency' }
        amountCents: { type: integer }
        refundedCents: { type: integer }
        providerRef: { type: [string, 'null'] }
        failureCode: { type: [string, 'null'] }
        paidAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
...
    Document:
      type: object
      required: [id, bookingId, docType, status, series, number, currency, totalCents, taxableCents, igvCents]
      properties:
        id: { type: string, format: uuid }
        bookingId: { type: string, format: uuid }
        paymentId: { type: [string, 'null'], format: uuid }
        docType: { $ref: '#/components/schemas/DocumentType' }
        status: { $ref: '#/components/schemas/DocumentStatus' }
        series: { type: string, example: B001 }
        number: { type: integer }
        currency: { $ref: '#/components/schemas/Currency' }
        totalCents: { type: integer }
        taxableCents: { type: integer, description: Base imponible. }
        igvCents: { type: integer }
        exchangeRate: { type: [string, 'null'], description: Tipo de cambio SUNAT si es en USD. }
        sunatCode: { type: [string, 'null'] }
        sunatMessage: { type: [string, 'null'] }
        relatedDocumentId: { type: [string, 'null'], format: uuid }
        hasXml: { type: boolean }
        hasCdr: { type: boolean }
        hasPdf: { type: boolean }
        issuedAt: { type: [string, 'null'], format: date-time }
        voidedAt: { type: [string, 'null'], format: date-time }

    # ---------------------------------------------------- Clientes y legal
    Customer:
      allOf:
        - $ref: '#/components/schemas/CustomerInput'
        - type: object
          required: [id]
          properties:
            id: { type: string, format: uuid }
            bookings:
              type: array
              items: { $ref: '#/components/schemas/BookingSummary' }
    Waiver:
      type: object
      required: [id, bookingId, status, version]
      properties:
        id: { type: string, format: uuid }
        bookingId: { type: string, format: uuid }
        passengerId: { type: [string, 'null'], format: uuid }
        version: { type: integer }
        status: { $ref: '#/components/schemas/WaiverStatus' }
        signerName: { type: [string, 'null'] }
        signedAt: { type: [string, 'null'], format: date-time }
        hasPdf: { type: boolean }
...